Refactor: split habit chart line and data processing
