+++ lib/widgets/charts/dashboard_habits_chart.dart
@@ -1,8 +1,6 @@
 import 'dart:core';
 
 import 'package:beamer/beamer.dart';
-import 'package:collection/collection.dart';
-import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:intersperse/intersperse.dart';
@@ -13,10 +11,10 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/themes/themes.dart';
 import 'package:lotti/utils/color.dart';
 import 'package:lotti/widgets/charts/dashboard_chart.dart';
-import 'package:lotti/widgets/charts/utils.dart';
+
+import 'dashboard_habits_data.dart';
 
 class DashboardHabitsChart extends StatefulWidget {
   const DashboardHabitsChart({
@@ -109,80 +107,6 @@ class _DashboardHabitsChartState extends State<DashboardHabitsChart> {
       },
     );
   }
-}
-
-class HabitResult extends Equatable {
-  const HabitResult(this.dayString, this.hexColor);
-
-  final String dayString;
-  final String hexColor;
-
-  @override
-  String toString() {
-    return '$dayString $hexColor}';
-  }
-
-  @override
-  List<Object?> get props => [dayString, hexColor];
-}
-
-final successColor = colorToCssHex(primaryColor);
-final failColor = colorToCssHex(alarm);
-final skipColor = colorToCssHex(
-  styleConfig().secondaryTextColor.withOpacity(0.4),
-);
-
-List<HabitResult> habitResultsByDay(
-  List<JournalEntity> entities, {
-  required HabitDefinition habitDefinition,
-  required DateTime rangeStart,
-  required DateTime rangeEnd,
-}) {
-  final resultsByDay = <String, String>{};
-  final range = rangeEnd.difference(rangeStart);
-  final dayStrings = List<String>.generate(range.inDays + 1, (days) {
-    final day = rangeStart.add(Duration(days: days));
-    return ymd(day);
-  });
-
-  final activeFrom = habitDefinition.activeFrom ?? DateTime(0);
-  final activeUntil = habitDefinition.activeUntil ?? DateTime(9999);
-
-  for (final dayString in dayStrings) {
-    final day = DateTime.parse(dayString);
-    final hexColor = (day.isAfter(activeFrom) || day == activeFrom) &&
-            day.isBefore(activeUntil)
-        ? failColor
-        : skipColor;
-
-    resultsByDay[dayString] = hexColor;
-  }
-
-  for (final entity in entities.sortedBy((entity) => entity.meta.dateFrom)) {
-    final dayString = ymd(entity.meta.dateFrom);
-    final hexColor = entity.maybeMap(
-      habitCompletion: (completion) {
-        final completionType = completion.data.completionType;
-        final hexColor = completionType == HabitCompletionType.fail
-            ? failColor
-            : completionType == HabitCompletionType.skip
-                ? skipColor
-                : successColor;
-
-        return hexColor;
-      },
-      orElse: () => skipColor,
-    );
-
-    resultsByDay[dayString] = hexColor;
-  }
-
-  final aggregated = <HabitResult>[];
-  for (final dayString in resultsByDay.keys.sorted()) {
-    aggregated.add(HabitResult(dayString, resultsByDay[dayString]!));
-  }
-
-  return aggregated;
 }
 
 class HabitChartInfoWidget extends StatelessWidget {
