+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: A Smart Journal.
 publish_to: 'none'
-version: 0.8.189+1555
+version: 0.8.190+1556
 
 msix_config:
   display_name: Lotti
